🐞 fix(FactorCompare.py): bugs fixed

+++ src/censo_ext/FactorCompare.py
@@ -142,7 +142,7 @@ def Factor_xyzCompare(args) -> None:
     diff2_Res: npt.NDArray[np.float64] = np.diff(np.diff(sort_Res))
     STD_diff2_Res: float = float(diff2_Res.std())
 
-    idx_max_diff2_R: npt.NDArray[np.float64] = np.array([], dtype=np.float64)
+    idx_max_diff2_R: npt.NDArray[np.int64] = np.array([], dtype=np.int64)
     for idx0, num in enumerate(diff2_Res):
         if num > STD_diff2_Res:
             idx_max_diff2_R = np.append(idx_max_diff2_R, idx0)
